added logo to portfolio section. it perfectly centers logo and scales it to fit the area.
