+++ src/App.css
@@ -2,3 +2,11 @@
   overflow: hidden;
   font-family: 'Open Sans', sans-serif;
 }
+
+h1, h2, h3, h4, h5, h6 {
+  color: #003459;
+}
+
+p {
+  color:#00171f; 
+}
