+++ src/Components/HeroCard/HeroCard.js
@@ -9,21 +9,31 @@ class HeroCard extends Component {
         return (
             
                 <div style={{textAlign: 'center'}}>
+
                     <div className="row">
+
                         <div className="animated fadeIn" style={{margin: '0 auto', zIndex: '2'}}>
-                        <img src={Logo} alt="logo" className="code-icon animated bounceInDown"/>
-                            <div className="intro-header animated fadeInDown">
+
+                            <img src={Logo} alt="logo" className="code-icon animated bounceInDown" />
+
+                                <div className="intro-header animated fadeInDown">
                                 
-                                <span className="intro-title">David Guben</span>
-                                <br/>Front-End Developer
-                            </div>
+                                    <span className="intro-title">David Guben</span>
+                                    <br/>Front-End Developer
+
+                                </div>
                         </div>
+
                     </div>
                     
                     <div className="row">
-                    <div className="intro-content animated fadeInLeft" style={{margin: '0 auto', zIndex: '1'}}>
+
+                        <div className="intro-content animated fadeInLeft" style={{margin: '0 auto', zIndex: '1'}}>
                         <h1><b>What I Do</b></h1>
-                        <p>Design & develop custom responsive websites or applications that look great on any device.</p>
+                        <p>
+                            Design & develop custom responsive websites or applications using JavaScript ES6, Node.JS, and React.JS. 
+                        </p>
+
                     </div>
                     </div>
 
@@ -35,7 +45,9 @@ class HeroCard extends Component {
                          heading="Designer"
                          tools="Adobe Create Suite: Photoshop, Illustrator, InDesign, Premiere"
                          icon="fa fa-pencil fa-4x">
-                            <p style={{textAlign: 'left'}}>I use Adobe Creative Suite to create custom designs and assets. I can help design anything to be used for your digital and print work.</p>
+                            <p style={{textAlign: 'left'}}>
+                                I use Adobe Creative Suite to create custom designs and assets. I can help design anything to be used for your digital and print work.
+                            </p>
                         </Service>
                                 
                         <Service
@@ -43,7 +55,9 @@ class HeroCard extends Component {
                          heading="Developer"
                          tools="HTML5, CSS3, JavaScript ES6, Node, React, Git, MongoDB"
                          icon="fa fa-code fa-4x">
-                            <p style={{textAlign: 'left'}}>I use JavaScript ES6 along with today's modern frameworks to create custom responsive web applications that work great on all kinds of devices.</p>
+                            <p style={{textAlign: 'left'}}>
+                                I use JavaScript ES6 along with today's modern frameworks to create custom responsive web applications that work great on all kinds of devices.
+                            </p>
                         </Service>
 
                     </div>
