--- src/Components/Portfolio/Portfolio.css
@@ -1,51 +1,0 @@
-.project-category {
-    box-shadow: 0px 0px 20px rgba(23, 23, 23, 0.3);
-    text-shadow: 1px 0px 0px rgba(43, 42, 42, 1);
-    position: relative;
-    font-size: 15px;
-    background-color: #007ea7;
-    border-radius: 20px;
-    padding: 5px;
-    width: 100px;
-    margin: 0 auto;
-    color: white;
-    top: -20px;
-    z-index: 1;
-  }
-
-  .folio-card {
-    border-radius: 20px;
-  }
-
-  .company-logo-container {
-    position: absolute;
-    width: 30px;
-    z-index: 1;
-    margin: 0 auto;
-  }
-
-  .company-logo {
-    width: 300px;
-  }
-
-  @media screen and (min-width: 0px) and (max-width: 991px) {
-    .company-logo {
-      width: 500px;
-    }
-  }
-  
-  @media screen and (min-width: 992px) and (max-width: 999px) {
-    .company-logo {
-      width: 220px;
-    }
-  }
-
-  @media screen and (min-width: 1000px) and (max-width: 1199px) {
-    .company-logo {
-      width: 240px;
-    }
-  }
-
-  @media screen and (min-width: 1200px) {
-
-  }
