+++ src/Components/ProjectCard/ProjectCard.css
@@ -1,0 +1,38 @@
+.project-category {
+    box-shadow: 0px 0px 20px rgba(23, 23, 23, 0.3);
+    text-shadow: 1px 0px 0px rgba(43, 42, 42, 1);
+    text-align: center;
+    position: relative;
+    font-size: 15px;
+    background-color: #007ea7;
+    border-radius: 20px;
+    padding: 5px;
+    width: 100px;
+    margin: 0 auto;
+    color: white;
+    top: -20px;
+    z-index: 1;
+  }
+
+  .folio-card {
+    border-radius: 20px 20px 20px 20px;
+    display: inline-block;
+    margin: 0 auto;
+  }
+
+  .company-logo-container {
+    position: absolute;
+    z-index: 1;
+    text-align: center;
+    width: 100%;
+  }
+  
+  .company-logo {
+    margin: 0 auto;
+    padding: 5px 5px 5px 5px;
+  }
+
+  .company-logo img {
+    width: 60%;
+    height: auto;
+  }
